+++ src/features/source/browse/screens/SourceMangas.tsx
@@ -60,6 +60,7 @@ import { VirtuosoUtil } from '@/lib/virtuoso/Virtuoso.util.tsx';
 import { IconWebView } from '@/assets/icons/IconWebView.tsx';
 import { MigrationManager } from '@/features/migration/MigrationManager.ts';
 import { ReactRouter } from '@/lib/react-router/ReactRouter.ts';
+import type { FilterChangeInput } from '@/lib/graphql/generated/graphql-base.types.ts';
 
 const DEFAULT_SOURCE: SourceIdInfo = { id: '-1' };
 
@@ -120,6 +121,7 @@ const useSourceManga = (
         GetSourceMangasFetchMutation,
         GetSourceMangasFetchMutationVariables
     >;
+
     switch (contentType) {
         case SourceContentType.POPULAR:
             result = requestManager.useGetSourcePopularMangas(sourceId, initialPages);
@@ -132,21 +134,27 @@ const useSourceManga = (
                 sourceId,
                 searchTerm ?? '',
                 filters.map((filter) => {
-                    const { position, state, group } = filter;
+                    const { positions, state } = filter;
 
-                    const isPartOfGroup = group !== undefined;
+                    const isPartOfGroup = positions.length > 1;
                     if (isPartOfGroup) {
-                        return {
-                            position: group,
-                            groupChange: {
+                        return positions.reduceRight((update, position, index) => {
+                            if (index === positions.length - 1) {
+                                return {
+                                    position: positions.slice(-1)[0],
+                                    [filter.type]: state,
+                                };
+                            }
+
+                            return {
                                 position,
-                                [filter.type]: state,
-                            },
-                        };
+                                groupChange: update,
+                            };
+                        }, {} as FilterChangeInput);
                     }
 
                     return {
-                        position,
+                        position: positions[0],
                         [filter.type]: state,
                     };
                 }),