+++ CHANGELOG.md
@@ -16,7 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Fixed
 
-- .
+- (**Browse**) Fix filtering with nested group changes
 
 ## [20260726.01] (r3379) - 2026-07-26